fix: Distribución de funciones del Wifi.
