+++ components/wifi_manager/CMakeLists.txt
@@ -1,0 +1,5 @@
+idf_component_register(
+    SRCS "wifi_manager.c"
+    INCLUDE_DIRS "include"
+    REQUIRES esp_wifi nvs_flash esp_netif esp_event lwip lvgl
+)
